Added in "Hard AI mode" and nerfed Autocracy NV
Added in the triggered modifiers from to help the ai manage it's tariffs, but they have to be enabled via a "Hard AI" additional option.
Nerfed the Autocracy National Value to help reduce Russian/Ottoman stability a bit.
Made Jingoism not matter for adding wargoals, doesn't seem to affect the AI much and it is better for multiplayer this way.