Foreign investment finally fixed and Ottoman changes
Finally after much trial and error made a fix for building railroads in other countries.
Also changed the ottoman start to not be at war in Libya, because (according to Wikipedia) their war ended there in 1835, and because it caused the ottomans to almost never actually annex them, so this Libyan rumpstate would often make it to 1936 and never be annexed/colonized.
Also made the Brazilian-Portuguese capital event make more sense for either country who founds the country.