Another Decision for Germany and balance changes for Autocracy and Absolute Monarchy
Added in a decision for Germany to remove the French cores in Alsace Lorraine in exchange for returning any other French cores Germany has taken.
Made the German birthrate reduction event trigger in 1875 instead of 1870, still trying to get the German population around what it was IRL in most games.
Also removed the war exhaust reduction for Autocracy to try and make Tzarist Russia a bit more likely to collapse lategame if they lose a big war
Now Absolute Monarchies can appoint Fascist parties to be the ruling party, this was done so that absolute monarchies can pass social reforms in the super late game, and because there is historical precedent for Monarchs to work with Fascists, usually against socialists/communists.