Changes to Germany, Russia, and France
Added a new additional rule to help isolate Russia, similar to the splendid isolation rule for the UK, which will prevent Russia from forming an alliance with anyone other than Austria and Prussia/NGF/Germany, who historically they did have alliances with. This was done because in most games Russia would ally France in the 1840's and participate in the Franco-Prussian wars, which often lead to it being occupied by Germany, and then rebel spiraling thereafter. Historically it also made sense, since Britain and France both deeply distrusted the Russian empire until the ~1890's, which isn't simulated in the game.
To help the Franco-Russian alliance form after 1890, France now has a decision to give Russia 1 million pounds to form an alliance.
Germany also commonly forms an alliance with Britain, which isn't implausible, but does happen more often than id like, so I added a decision for Germany to fund a naval expansion, like they did IRL from 1898-1914. This was a direct challenge to Britain, and resulted in the Alienation of Germany from the UK. Ingame this decision will allow Germany, for the price of 1 million pounds, to gain faster ship construction for 10 years. This will cause the UK to dislike Germany, and break an alliance if they have one together when Germany enacts this decision.