Fixed Hungarian Revolution and Taiping Rebellion
Hungary should now be instantly inherited, instead of being a vassal for a day, so AI Austria wont release Dalmatia.
Found the reason the Taiping have been winning so easily most games, so i removed the mobilization size they got from the modifier, because Uncivs aren't supposed to be able to mobilize, but could because this modifier stacked with the "save the nation" modifiers.